pyproject.toml
[project]
name = "audio-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "lightning>=2.5.6",
    "pre-commit>=4.4.0",
    "rich>=14.2.0",
    "rootutils>=1.0.7",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchmetrics>=1.8.2",
    "wandb>=0.23.0",
    "h5py>=3.10.0",
    "pandas>=2.2.0",
    "einops>=0.7.0",
    "timm>=0.9.0",
    "matplotlib>=3.10.7",
    "soundfile>=0.13.1",
]
